projects/eco_system_v2/requirements.txt
anthropic>=0.40.0
httpx>=0.27.0
yfinance>=0.2.40
fredapi>=0.5.2
python-dotenv>=1.0.0
